Fix searches that contain certain special symbols.
When performing a search from the home page it was encoding the search
query into the URL path in the form of "search/q/{query}". But this fails
for many different symbols. Even if you try to URI encode the query a "/"
will still break it. This commit changes the URL to be
"search?search_form=get_results&q={query}" and URI encodes the query. This
seems to work fine for every possible query with special symbols I could
think of. The catalog code also had a similar bug and another issue that
broke some special symbols.